------------------------------------------------------------------------ r3105 | horda03 | 2010-06-04 14:15:13 +0100 (Fri, 04 Jun 2010) | 70 lines
Cross integration of change 505253

On VMS runing the "asjava" test suite reproduces a
hange during the selv2 test phase.

Analysis of the hang revealed that the session had just
initialted an QIO (GCsend) operation, it was then in the
process of advancing to the next GCA state of GC_ASYNC
(introduced in fix for bug 121341).

When the QIO operation completes, an AST is executed which
executes a GCsend_complete. An AST interrupts the "normal"
process (on VMS that's the current active session), does its
processing then the next AST is executed )if one exists) otherwise
the normal session resumses. If the "normal" session is the one that
initiated the QIO, then if it is in gca_service where it is calculating
the next "state" to execute, the AST will perform the "current" state
(GC_ASYNC) and increment the state to the next one.

Alas, the "normal" session had already determined that the state to
execute was the GC_ASYNC, but hadn't managed to update the state value
in memory - because an operation like "val++" is typically a

read val
increment val
store val

(and the number of operations between the READ and STORE could increase with
optimisation).

On VMS you can't use an Ingres mutex to protect the update against access by an
AST, because should the AST be prevented from obtaining the mutex the system will
hang. VMS (like all other platforms) provide native Atomic instructions. An atominc
instruction is one that must complete without any "interference" from another
process, in an increment say, if the Atomic increment fails, the value will be unchanged,
and the increment retried.

Because of the overhead of an Atomic increment (we believe small but more than a simple ++)
the atomic increment will be limited to VMS platforms. A new package within the GL has been
created. Currently this package consists of the header file at.h, but the intension is to
migrate the native platform atomic commands to it (for example CStas, CScas, CSadjust_counter)
for use. Currently the at.h command only defines the VMS atomic increment command for an I4
for use in updating variables which could be affected by AST executed code, on non-VMS
platforms this will revert to a simple ++ operation.

gcaasync.c

Added a local variable state which is used to store the current value of
svc_parms->state before it is incremented. This variable was added so that the
trace information displayed will be consistent (i.e. an AST coming along
which increments the state will not cause misleading information being
displayed.

The new macro GCA_INCREMENT_STATE is used to increment svc_parms->statei.

at.h

New header file for Atomic functions, currently only defines AT_ATOMIC_INCREMENT_I4
if this macro is used on any platform other than VMS it will cause a build failure.
It is outside the scope of this chage to implement atomic functions across all
platforms.

gc.h

On VMS define GCA_ATOMIC_INCREMENT_STATE

Define GCA_INCREMENT_STATE to AT_ATOMIC_INCREMENT_I4 if GCA_GCA_ATOMIC_INCREMENT_STATE defined,
otherise it is defined as the (non-atomic) ++ operation.

------------------------------------------------------------------------
r3106 | hanal04 | 2010-06-04 17:15:15 +0100 (Fri, 04 Jun 2010) | 21 lines

Cross integration of change 505268

The -add_drop option is used to generate a copy.in script that
can be re-run because it contains the necessary drop statements
before any create table statements.
If the table contains a column that makes use of a sequence the
CREATE SEQUENCE statements are generated but we fail to generate
the DROP SEQUENCE statements that allow the script to be re-run.

We now call xfdrop_sequences() if XF_DROP_INCLUDE has been specified
and XF_NO_SEQUENCES is not set.

xfdrop_sequences() has been updated so that it will only generate
DROP SEQUENCE statements for all tables if an Obj_list has not
been specified. If an Obj_list has been specified only those
sequences that relate to the specified tables will be generated.
The new code in xfdrop_sequences() is based on the existing
Obj_list code in xfsequences().

------------------------------------------------------------------------