------------------------------------------------------------------------ r3106 | hanal04 | 2010-06-04 17:15:15 +0100 (Fri, 04 Jun 2010) | 21 lines
Cross integration of change 505268

The -add_drop option is used to generate a copy.in script that
can be re-run because it contains the necessary drop statements
before any create table statements.
If the table contains a column that makes use of a sequence the
CREATE SEQUENCE statements are generated but we fail to generate
the DROP SEQUENCE statements that allow the script to be re-run.

We now call xfdrop_sequences() if XF_DROP_INCLUDE has been specified
and XF_NO_SEQUENCES is not set.

xfdrop_sequences() has been updated so that it will only generate
DROP SEQUENCE statements for all tables if an Obj_list has not
been specified. If an Obj_list has been specified only those
sequences that relate to the specified tables will be generated.
The new code in xfdrop_sequences() is based on the existing
Obj_list code in xfsequences().

------------------------------------------------------------------------
r3107 | abbjo03 | 2010-06-05 02:30:09 +0100 (Sat, 05 Jun 2010) | 7 lines

Correct change 504118 (cross-integration of 502560).

Phase 1 of single byte optimizations.

New data members should be added at end of clfelib.lot.

------------------------------------------------------------------------