------------------------------------------------------------------------ r3113 | hanal04 | 2010-06-07 11:15:10 +0100 (Mon, 07 Jun 2010) | 9 lines
When iiimport updates objects it does not mark them as
requiring recompilation (APC_RECOMPILE). This change
updates IIIEwao_WriteIIAbfObjects() to add the
APC_RECOMPILE to the abf_flags value for all objects
that are written to ii_abfobjects. This ensures a
subsequent iiimageapp will recompile the updated
objects.

------------------------------------------------------------------------
r3114 | horda03 | 2010-06-07 14:00:11 +0100 (Mon, 07 Jun 2010) | 39 lines

Cross integration of change 505273

VDBA does not allow a GRANT on a SEQUENCE to be dropped.

The SQL that VDBA generates to REVOKE the GRANT is
incorrect.

Currently VDBA generates

revoke on sequence <seq_name> from group <group_name>

But correct syntax is

revoke NEXT on sequence <seq_name> from <user> <cascade|restrict>

rvktable:

Don't hide the CASCADE/RESTRICT buttons for revoking a GRANT on a
sequence. Also, make CASCADE the default.

Removed duplicate Button_SetCheck() call for TABLE_PRIVILEGE

dbasqlpr.c

For a REVOKE of a GRANT on a SEQUENCE add " next " to the SQL
string and CASCADE or RESTRICT (as appropriate).

dom2.c

Now that the correct syntax the DROP grant on a sequence is being
generated, prevent message

Drop not implemented yet for this type of object.

being displayed (even though GRANT has been removed. Add the
OT_SEQUGRANT_NEXT_USER case.

------------------------------------------------------------------------