------------------------------------------------------------------------ r3122 | drivi01 | 2010-06-08 17:15:12 -0400 (Tue, 08 Jun 2010) | 3 lines
Fix NT_GENERIC spelling for DEBUG to avoid inclusion of strings.h.

------------------------------------------------------------------------
r3123 | maspa05 | 2010-06-09 05:30:11 -0400 (Wed, 09 Jun 2010) | 11 lines

\vdelim NONE leads to missing lines of output when run under SEP.

\vdelim NONE sets the vertical separator character, drchtbl[DRQK].prchar, to
EOS. Whilst this works for normal operation because the output routines use
the length of the buffer, SEP communicates with child processes via files
and when 'reading' the output it sees the EOS in the string and thinks that's
the end of the line.

Change qrputc to not add EOS's into the buffer.

------------------------------------------------------------------------