------------------------------------------------------------------------ r3123 | maspa05 | 2010-06-09 05:30:11 -0400 (Wed, 09 Jun 2010) | 11 lines
\vdelim NONE leads to missing lines of output when run under SEP.

\vdelim NONE sets the vertical separator character, drchtbl[DRQK].prchar, to
EOS. Whilst this works for normal operation because the output routines use
the length of the buffer, SEP communicates with child processes via files
and when 'reading' the output it sees the EOS in the string and thinks that's
the end of the line.

Change qrputc to not add EOS's into the buffer.

------------------------------------------------------------------------
r3124 | horda03 | 2010-06-09 08:00:11 -0400 (Wed, 09 Jun 2010) | 4 lines

Fix adf.h to allow builds with xDEBUG defined
to work.

------------------------------------------------------------------------