------------------------------------------------------------------------ r3127 | drivi01 | 2010-06-09 17:15:12 -0400 (Wed, 09 Jun 2010) | 11 lines
Ingres VectorWise table icon was appearing for non-VectorWise tables from time
to time.

This is due to the fact that the memory of the structure where table
information is stored is never cleared out, this was causing leftover
garbage in the memory to be picked up as IVW table.

Also, updated icon for tables in the properties pain to reflect IVW tables
when static "Tables" item is selected in the tree.

------------------------------------------------------------------------