------------------------------------------------------------------------ r3130 | frima01 | 2010-06-10 05:45:15 -0400 (Thu, 10 Jun 2010) | 4 lines
Correct rel_v9.reldatawid and rel_v9.reltotdatawid for upgrade case.
Have iistats updated for Ingres 10 distributed databases.

------------------------------------------------------------------------
r3131 | coomi01 | 2010-06-10 09:15:16 -0400 (Thu, 10 Jun 2010) | 27 lines

This piece of SQL throws an error, due to combined presence
of distinct and sequence in the query viewed as a whole:

	select
		MYSEQ.nextval, user_name
	from
		users

	where   user_name not in
	(
		select  DISTINCT
			username
		from
			banned_Users
	)

Use the Ysave stack to push fresh value of seq_op onto the stack
on entry to a portion of sub-select, and pop it on return,
effecting a 'scope of' seq_op action.

Test the stack rather than global value before generating the E_US18AF.

Importantly: Retain older global test this is because for some constructs,
creation of views for example, the use of seq_ops is forbidden no matter
the what the scope is.

------------------------------------------------------------------------