------------------------------------------------------------------------ r3134 | ingres | 2010-06-11 05:15:48 -0400 (Fri, 11 Jun 2010) | 3 lines
Created the file in error

------------------------------------------------------------------------
r3135 | horda03 | 2010-06-11 05:15:54 -0400 (Fri, 11 Jun 2010) | 20 lines

Cross integration of change 505324

In VDBA the commands entered are parsed and individual
commands are sent to the DBMS server.

The ';' is the comamnd delimiter, but in the case of a
CREATE PROCEDURE comamnd, the ';' should be ignored
in the DECLARE..BEGIN..END sections.

VBDA handled the above case, even with nested BEGIN..END
section. Now however, Ingres allows CASE..END sections.
If these are used, then VDBA broke up the comamnd too
early, leading to invalid queries.

This change keeps track of the BEGIN and CASE statements,
and only when the number of these match the number of
END commands does it allow a ';' to delimit a command.

------------------------------------------------------------------------