------------------------------------------------------------------------ r3135 | horda03 | 2010-06-11 05:15:54 -0400 (Fri, 11 Jun 2010) | 20 lines
Cross integration of change 505324

In VDBA the commands entered are parsed and individual
commands are sent to the DBMS server.

The ';' is the comamnd delimiter, but in the case of a
CREATE PROCEDURE comamnd, the ';' should be ignored
in the DECLARE..BEGIN..END sections.

VBDA handled the above case, even with nested BEGIN..END
section. Now however, Ingres allows CASE..END sections.
If these are used, then VDBA broke up the comamnd too
early, leading to invalid queries.

This change keeps track of the BEGIN and CASE statements,
and only when the number of these match the number of
END commands does it allow a ';' to delimit a command.

------------------------------------------------------------------------
r3136 | horda03 | 2010-06-11 06:00:10 -0400 (Fri, 11 Jun 2010) | 19 lines

Cross integration of change 505338

The command

ifnull( decimal_col, char_col)

fails with E_OP0791_ADE_INSTRGEN if the char_col is nullable.

In opc_adtransform the length of the resultant datatype is
calculated from the datatype's precision this however does
not include the extra byte needed for a nullable datatype.

Later when the decimal datatype is validated, 1 is subtracted
from the length  for comparison with the decimals length
(based on precision) this was causing the E_OP0791_ADE_INSTRGEN
error.

------------------------------------------------------------------------