------------------------------------------------------------------------ r3136 | horda03 | 2010-06-11 06:00:10 -0400 (Fri, 11 Jun 2010) | 19 lines
Cross integration of change 505338

The command

ifnull( decimal_col, char_col)

fails with E_OP0791_ADE_INSTRGEN if the char_col is nullable.

In opc_adtransform the length of the resultant datatype is
calculated from the datatype's precision this however does
not include the extra byte needed for a nullable datatype.

Later when the decimal datatype is validated, 1 is subtracted
from the length  for comparison with the decimals length
(based on precision) this was causing the E_OP0791_ADE_INSTRGEN
error.

------------------------------------------------------------------------
r3137 | jenjo02 | 2010-06-11 15:30:17 -0400 (Fri, 11 Jun 2010) | 6 lines

SIR 121123 Long Names

Sensitize hash of LK_CREATE_TABLE lock key table owner and table name
to long names.

------------------------------------------------------------------------