------------------------------------------------------------------------ r3140 | jenjo02 | 2010-06-14 11:30:11 -0400 (Mon, 14 Jun 2010) | 33 lines
SD 134198, Bug 123896

Sequences and/or identity columns with names that are not unique in the
first 12 characters may produce inconsistent sequence values.

The function dms_open_seq() uses the sequence owner and sequence name
to compose a LK_SEQUENCE lock key.

The first 8 bytes of the owner name are copied to lk_key2 and lk_key3,
and the first 12 bytes of the sequence name are copied to lk_key4,
lk_key5, and lk_key6.

If the resulting lock resource key is not unique, the same LK_SEQUENCE
lock will be used for differently owned or named sequences and may cause
subsequent operations on the sequence to make incorrect assumptions
about its state.

The fix is for qeu_seq to acquire the sequence id before opening
it in DMF. This assures that a sequence id will always be available
to dms_open_seq(), and a unique LK_SEQUENCE lock key can be formed
from database id and sequence id instead of owner and name.

Added LK_SEQUENCE display to dmd_lock for debugging when lock_trace
is turned on, and a call to it from dms_open_seq().

Also added a distinct display of LK_SEQUENCE locks to LKkey_to_string().

Propagation to other codelines:

I will cross the change to ingres93, applicable to all earlier codelines
with sequence generator support.

------------------------------------------------------------------------
r3141 | jenjo02 | 2010-06-14 15:00:12 -0400 (Mon, 14 Jun 2010) | 19 lines

SD 121619 MVCC:

dmd_check(E_DMF011_DM1R_BAD_TID) due to transaction's attempt
to delete a row it had previously deleted.

Caused by the transaction finding the pre-delete CR page already
in cache.

Now in dm0pUnmutex(), if the transaction unmutexing the Root
has changed the page, force rematerialization of any CR pages
constructed for the transaction with serializable isolation;
page changes must be visible to serializable transactions.

Also noticed that the computation of svcb_xid_lg_id_max
(the largest possible log_id) by dmcstart was incorrect,
affecting only the dmp.h macro IsTranActive and only if
a very large number of concurrent transactions have been run.

------------------------------------------------------------------------