------------------------------------------------------------------------ r3143 | gupsh01 | 2010-06-14 16:00:11 -0400 (Mon, 14 Jun 2010) | 14 lines
On UTF8 installations, extra long varchar string constants
(~16000 or more) inserts into nchar/nvarchar columns fails
with E_PS037B_UNORM error. This was observed when running
sep test dtu49.sep. The error was being returned when
Varchar string constant data was normalized before inserts
and the length of input varchar string constant was found
more than the maximum length allowed in a UTF8
installation.

The input string is truncated to the maximum allowed length.
If the connection option string_truncation=fail is supplied,
then an error will be returned to the user.

------------------------------------------------------------------------
r3144 | hanal04 | 2010-06-15 04:00:12 -0400 (Tue, 15 Jun 2010) | 9 lines

Cross integration of change 505320

Added a %^ option to TRdisplay so that the pid,tid of the caller can
be easily generated.
In VMS the tid will always be the same number until we get threading
implemented.

------------------------------------------------------------------------