------------------------------------------------------------------------ r3144 | hanal04 | 2010-06-15 04:00:12 -0400 (Tue, 15 Jun 2010) | 9 lines
Cross integration of change 505320

Added a %^ option to TRdisplay so that the pid,tid of the caller can
be easily generated.
In VMS the tid will always be the same number until we get threading
implemented.

------------------------------------------------------------------------
r3145 | hanal04 | 2010-06-15 04:00:27 -0400 (Tue, 15 Jun 2010) | 9 lines

Cross integration of change 505367

Rewrite fix for Bug 117368 which breaks optimizedb when run against
a Star Database. iirelation does not exist in a Star Database. We do
not need iirelation to exclude Gateway tables from the select.
iitables.table_subtype will equal 'I' for a gateway table.

------------------------------------------------------------------------