------------------------------------------------------------------------ r3145 | hanal04 | 2010-06-15 04:00:27 -0400 (Tue, 15 Jun 2010) | 9 lines
Cross integration of change 505367

Rewrite fix for Bug 117368 which breaks optimizedb when run against
a Star Database. iirelation does not exist in a Star Database. We do
not need iirelation to exclude Gateway tables from the select.
iitables.table_subtype will equal 'I' for a gateway table.

------------------------------------------------------------------------
r3146 | schka01 | 2010-06-15 06:45:11 -0400 (Tue, 15 Jun 2010) | 6 lines

Fix "deprecated routine" warnings and build problems
on Mac OSX.  In particular, CMcpychar was being used
with a constant string, which fails when the constant
is compiled into the very end of the __const section.

------------------------------------------------------------------------