------------------------------------------------------------------------ r3146 | schka01 | 2010-06-15 06:45:11 -0400 (Tue, 15 Jun 2010) | 6 lines
Fix "deprecated routine" warnings and build problems
on Mac OSX.  In particular, CMcpychar was being used
with a constant string, which fails when the constant
is compiled into the very end of the __const section.

------------------------------------------------------------------------
r3147 | horda03 | 2010-06-15 07:30:10 -0400 (Tue, 15 Jun 2010) | 4 lines

Change dependancy to the VERSFILE symbol rather than the VERS and
VERS.<platform> file. No need for the NOCARE now.

------------------------------------------------------------------------