------------------------------------------------------------------------ r3151 | horda03 | 2010-06-16 09:15:10 -0400 (Wed, 16 Jun 2010) | 30 lines
Cross integration of change 505419

The nightly build tests of Ingres 10 unearthed a hang running the
STAR tests. By a process of elimination, the fix for Bug 123840
has been identified as the cause of the hang.

The purpose of this fix was to prevent a hang on VMS, where AST code
was updating the GCA's state variable "at the same time" as the main
code. This causes a "corruption" in the state engine processing for the
session, and lead to the session hanging.

At the time the change was made, it was felt that only VMS required
an ATOMIC instruction to increment the state variable (a mutex is
not approprate as using a mutex to protect code from an AST will lead
to a hang). It was also felt that many areas within Ingres could benefit
from the availability of atomic functions; but this was not the time
to begin such a review; as atomic functions are platform specific.

This change defines an Atomic increment function for linux, the atomic.h
file supplied on linux (/usr/include/atomic.h) doesn't provide a suitable
function for our needs, hence the function is written in machine code.

NOTE: Without this fix the nightly build on a64.lnx hang 100% of the time in
the STAR tests. This implies that two threads are updating the same svc_parms's
state variable. If this is true on one linux/unix platform, then its possible that
the same situation can happen on others (this includes NT - VMS already has
the fix).

------------------------------------------------------------------------