------------------------------------------------------------------------ r3153 | schka01 | 2010-06-16 10:45:10 -0400 (Wed, 16 Jun 2010) | 9 lines
Parse tree nodes are allocated exact-size, and
most are smaller than the entire PST_QNODE.  Fix
code that copied the entire PST_QNODE from a
passed-in parse tree pointer, thus addressing past
the end of the pointed to node.  If the node
happened to be at the end of the address space, a
segv occurred.

------------------------------------------------------------------------