------------------------------------------------------------------------ r3153 | schka01 | 2010-06-16 10:45:10 -0400 (Wed, 16 Jun 2010) | 9 lines
Parse tree nodes are allocated exact-size, and
most are smaller than the entire PST_QNODE.  Fix
code that copied the entire PST_QNODE from a
passed-in parse tree pointer, thus addressing past
the end of the pointed to node.  If the node
happened to be at the end of the address space, a
segv occurred.

------------------------------------------------------------------------
r3154 | loera01 | 2010-06-16 13:30:10 -0400 (Wed, 16 Jun 2010) | 10 lines

Cross integration of change 505328 from ingres2006r3.

Examination of ConfigDSN(), which is the fuction executed for ODBC batch
configuration, showed a number of inaccuracies when copying and normalizing
the input arguments to internal buffers.  Some general code clean-up was
requires as well, since the code was nearly illegible due to being edited
originally in DevStudio.  There was also places where functions as strcmp()
and memcmp() wereused instead of their compatlib counterparts.

------------------------------------------------------------------------