------------------------------------------------------------------------ r3157 | kiria01 | 2010-06-16 18:08:57 -0400 (Wed, 16 Jun 2010) | 14 lines
Fix b123908.

ULM memory requests to streams were sometimes passing uninitialised
ulm_rcb data resulting on sporadic SEGVs. Similarly, pss_stk_freelist
is occasionally not reset after a stream was wound back.

Description of change:

ULM control blocks have a pointer, .ulm_streamid_p that should either point to a valid address into which to store the stream id or should be NULL. As this block is often allocated on the stack, not initialising this can cause segvs or misinterpretation of data. This is not initialised to NULL when not otherwise set. Also noted was an instance of uninitialised ulm_flags setting which is now corrected.

The parser session block member .pss_stk_freelist caches a list of block for reuse but this must be NULL when the psf_mopen call has been done. This is now cleared.

A further access error was corrected by checking that the PST_QNODE involved was the correct type before looking at the data.

------------------------------------------------------------------------