------------------------------------------------------------------------ r3162 | stephenb | 2010-06-17 21:17:56 -0400 (Thu, 17 Jun 2010) | 47 lines
Release Notes:
rns.start

     (GENERIC)
     123880
     The nvl2 SQL function always returns a "char" type and a length based on parameters 1 and 2
     no matter what types are passed in. The correct behaviour is to return a type and length
     based on parameters 2 and 3.

     (GENERIC)
     123884
     the nvl2 SQL function will return error E_SC0206 when non compatible types are passed;
     sometimes the session is terminated. The function should return a user error indicating
     the actual problem.
rns.end

Description of changes:
desc.start
The FI table description for NVL2 was corrected to return DB_ALL_TYPE, and a new lenspec,
ADI_LONGER23RES, was created to cope with the special length needs of NVL2. Since all
parameters and the return type are now DB_ALL_TYPE, and no directive on return type is
hard coded into the FI table, special code is needed to determine the return type at
runtime. This is intentional since the types of parameters 2 and 3 should be used to
determine the correct return type, this is unique (currently) to NVL2. New code
was added to support this runtime determined return type using the standard typing
rules defined in adi_resolve. New code was also added to correctly calculate the
length of the return based on the lengths of parameter 2 and parameter 3 (again unique
to NVL2) using the directive, ADI_LINGER23RES, from the FI table.

When incompatible types or incorrect parameters are passed to NVL2,
ADF functions will return E_DB_ERROR with a specific ADF error to indicate
the source of the problem. PSF usually handles
these specific ADF errors and converts them to user errors that are ultimately returned
to the user. In the absence of this special error handling code in PSF, the user would
otherwise see a error or severe error return status from PSF with no other explanation,
this results in nasty SCF errors in the front-end and sometimes causes the session to
terminate to due to exceeding "max error for the session". Some ADF function calls made by
PSF when processing NVL2 did not have the correct error handling to cope with the possible
user errors generated by bad types or insufficient parameters, thus causing
nasty errors and session terminations. This error handling
was added to the code.

Because NVL is an alias for ifnull, some errors display "ifnull" as the function in error, a
minor alteration to error messaging was coded to indicate that it could be ifnull or nvl.
desc.end

------------------------------------------------------------------------
r3163 | stial01 | 2010-06-18 16:00:21 -0400 (Fri, 18 Jun 2010) | 10 lines

SIR 121619 MVCC

Description of changes:
Added more TRdisplays for btree reposition errors.
In dmddata fixed trace output for compressed keys, added dmd_pr_mvcc_info,
deleted dmdprbkey()
In dm1cidx.c added dm1cx_txn_get to get entry transaction information

------------------------------------------------------------------------