------------------------------------------------------------------------ r3165 | drivi01 | 2010-06-21 17:15:16 -0400 (Mon, 21 Jun 2010) | 5 lines
This is a Release profile for a visual C project for change #505498.
The change requires some additional headers, and only debug version had an
updated profile.

------------------------------------------------------------------------
r3166 | coomi01 | 2010-06-22 08:15:09 -0400 (Tue, 22 Jun 2010) | 9 lines

The unextend-db utility does not check on the number
of work locations before deleting the final one.

Before calling the stored proc to do the work, quickly
count the work locations for each database being manipulated.
If the count is 1, do not allow the deletion to go ahead,
producing an error messiage instead.

------------------------------------------------------------------------