------------------------------------------------------------------------ r3175 | maspa05 | 2010-06-23 10:45:16 -0400 (Wed, 23 Jun 2010) | 31 lines
Cross integration of change 505528

In VDBA if you connect to a vnode, click on a database, then a table and then
choose the 'rows' tab you will create two connections to the database. One of
these is used by VDBA for meta-data etc. The other is opened for the actual
rows data. The "rows" pane creates its connection by invoking the sqlquery.ocx
(Visual SQL control). If this window is closed the connection remains open.

Amended the CuDlgDomPropTableRows::OnDestroy method to call the SqlQuery
GetConnected method. This checks the connections and returns 1 or 0 depending
on whether there are active connections or not. However it also disconnects
any sessions marked as "released". So the practical upshot of calling it here
is to close the connection for the rows pane.

Note that this closes all such connections including those used by any other
panes connected to the same vnode. However such connections are marked
"released" and they are holding no locks i.e. they are not in use. If after
the connection has been disconnected the pane needs to use it, it will simply
make a new connection. Thus it is safe to disconnect all "released" sessions.

Also change CSqlqueryCtrl::GetConnected to check for lpszDatabase being an
empty string as well as NULL because this parameter is passed as a VTS_BSTR
where a NULL can get treated as (i.e. converted to) an empty string. In our
own code we only ever call GetConnected with lpszDatabase=NULL (in
ExistSqlQueryAndIpm, called from execrmcmd1). If someone is calling GetConnected
with an empty string relying on the fact that it always returns 0 then they
already may not be getting expected behaviour depending on which compiler they
use (i.e. whether NULL gets converted to "").

------------------------------------------------------------------------
r3176 | shust01 | 2010-06-23 12:15:09 -0400 (Wed, 23 Jun 2010) | 5 lines

Cross integration of 2006r3 change 505478.
Include header file cv.h so #define for CVupper is picked up when compiling.
Needed as part of performance fix for bug b123139.

------------------------------------------------------------------------