------------------------------------------------------------------------ r3181 | grant | 2010-06-24 13:06:50 -0400 (Thu, 24 Jun 2010) | 12 lines
Trac #624 - fasti fails to detect the windows timezone name in Vista / Windows 7 / 2008 Server

The gettimezone program takes the Windows timezone string and tries
to match it with an equivalent Ingres II_TIMEZONE_NAME value. From
Windows Vista onwards the registry key used to store the timezone
string has changed from "StandardName" to "TimeZoneKeyName".

This change adds a version check to gettimezone.exe using
GetVersionEx(). If OSVERSIONINFO.dwMajorVersion >=6 the new key name
will be used.

------------------------------------------------------------------------
r3182 | lunbr01 | 2010-06-24 14:00:25 -0400 (Thu, 24 Jun 2010) | 9 lines

Initialize dwBytes stack variable to zero in function
GCwinsock2_OP_RECV_complete() so that if a WSARecv() must be
reissued to get the rest of the message, then the trace won't
display a bogus value when the I/O does not complete right away.
Also added a few trace statements and clarified some others in
this section of code to make it more obvious (in the trace) that
another WSARecv() had to be reissued and why.

------------------------------------------------------------------------