------------------------------------------------------------------------ r3183 | schka01 | 2010-06-25 11:15:32 -0400 (Fri, 25 Jun 2010) | 41 lines
Revamp the validation step for table procedures.
Previously, table procedures were validated (QP checked
against table timestamps) at each calling-QP action.
This was unnecessary work, and the code was insufficiently
careful to validate against the proper resource list
using the proper DSH.  In some cases this led to
queries terminating prematurely, apparently without error,
because tproc validation had gone astray.

What is done now is to validate table procedures along
with tables, when the first action of a QP is validated.
Each table procedure called by the QP is loaded and a
DSH for that tproc is created;  then, the tproc QP is
validated recursively.  If any tproc is invalid, or
needs to be recompiled, validation returns to the
sequencer to recreate the tproc QP.  This is a more
orderly process and keeps all the lists and DSH's
straight.  There is no longer any need for
"valid lists" for table procedures, just a
resource list entry.  The validation DSH is used
by the first QEN_TPROC node (for that tproc) that
is executed;  if there are additional references to
the same tproc in the QP, additional DSH's are
created as needed.

Cleanup of queries using table procedures had various
problems: bad lock counting (leading to tproc QP's
being "stuck" in QSF), improper DSH cleanup, improper
handling of tproc recreate failures.  These have all
been fixed.

Additionally, the tracking for table validation has
been changed.  Instead of using a very confusing and
space-wasting list of opened valid-list entries,
we simply check the related DMT_CB to see if the
corresponding table is open.  The resource list
and valid list entries were pruned to get rid of
redundancy and unused garbage.

------------------------------------------------------------------------
r3184 | drivi01 | 2010-06-25 11:30:13 -0400 (Fri, 25 Jun 2010) | 12 lines

It appears that all the binaries built by Ingres are copyrighted
with year 2004.

To provide copyright year and build information we link gv.rc with
all the binaries.  However, gv.rc has a copyright year hardcoded
to 2004.

This change updates gv.rc.template to replace copyright year
with a variable and updates mkgv.bat to update the variable
with the current year.

------------------------------------------------------------------------