------------------------------------------------------------------------ r3184 | drivi01 | 2010-06-25 11:30:13 -0400 (Fri, 25 Jun 2010) | 12 lines
It appears that all the binaries built by Ingres are copyrighted
with year 2004.

To provide copyright year and build information we link gv.rc with
all the binaries.  However, gv.rc has a copyright year hardcoded
to 2004.

This change updates gv.rc.template to replace copyright year
with a variable and updates mkgv.bat to update the variable
with the current year.

------------------------------------------------------------------------
r3185 | drivi01 | 2010-06-25 12:00:16 -0400 (Fri, 25 Jun 2010) | 13 lines

There's several small problems in Visual DBA related to long id changes:
1. Drop of tables with long ids results in a crash, but the table is dropped.
2. Rows property tab only displays 63 characters in the column names.
3. Pages property throws an error.
4. IIA/IJA on import/export display truncated column names.

This change fixes all these issues and a few others.
This change also fixes buffer limitation in SQL Scratchpad.  I discovered that
SQL Scratchpad is unable to handle very long blob columns or character columns
due to buffer limitation of 1000, i've updated the code to allocate the buffer
on the fly if 1000 is too small.

------------------------------------------------------------------------