------------------------------------------------------------------------ r3185 | drivi01 | 2010-06-25 12:00:16 -0400 (Fri, 25 Jun 2010) | 13 lines
There's several small problems in Visual DBA related to long id changes:
1. Drop of tables with long ids results in a crash, but the table is dropped.
2. Rows property tab only displays 63 characters in the column names.
3. Pages property throws an error.
4. IIA/IJA on import/export display truncated column names.

This change fixes all these issues and a few others.
This change also fixes buffer limitation in SQL Scratchpad.  I discovered that
SQL Scratchpad is unable to handle very long blob columns or character columns
due to buffer limitation of 1000, i've updated the code to allocate the buffer
on the fly if 1000 is too small.

------------------------------------------------------------------------
r3186 | horda03 | 2010-06-25 13:15:11 -0400 (Fri, 25 Jun 2010) | 89 lines

Cross integration of change 505604

On VMS the Logging/Locking shared memory section is a
Global Section that is deleted when the last session
connected to it drops its connection.

Typically this means when the DMFRCP is stopped the
lglkdata section is released. However, it is possible
for an IPM session to be connected to the lglkdata, if
this is the case then the memory will not be released,
and may still exist when the installation is next started.

The first fix for Bug 122555 handled the above situation
when Ingres was started, INGSTART detects that a stale
lglkdata memory section exists, and defines a logical
(GBL$II_<inst_id>_lglkdata.mem) so that a new lglkdata
memory section will be created.

This change extends the above fix, so that TX log files
can be initialised during an upgrade. If the upgrade
uses a differnet structure for the Logging/Logging system, then
attaching to the memory created by the older version will
cause unexpected errors.

iimkcluster.dcl

Need to know the II_INSTALLATION id, so ensure the value
is obtained (if it exists)

Use check_lglk_mem function (see below) to check the
logging/Locking memory and Online status.

check_lglk_mem returning ii_status_warning implies the
Logging system is online.

iishlib.dcl

Add definition of check_lglk_mem function.

<fixed typo>

check_lglk_mem will return ii_status_ok      if Logging System Offline
ii_status_warning if Online

Define the name of the logical which is used to use an Alternate
lglkdata global memory section.

Do the online check using the memory section that would be in use
(i.e. if alternate area in use, then use that).

Return warning if the Logging system is online.

If the logical name was defined (an system is offline) then extract
the pid from the logical value, and deassign the logical.

Now check if anything is connected to the "default" lglkdata global memory section.
If there is, then define the logical. Note, if the logical was previously
defined, then add 1 to the PID obtained to make a unique number, otherwise
use the PID of the current process.

Return OK as the Logging system is offline.

iisudbms.dcl

Use check_lglk_mem to ensure a new lglkdata global memory section will
be used to create/upgrade the TX log files.

iisunode.dcl

Need to know the II_INSTALLATION id, so ensure the value
is obtained (if it exists)

Use check_lglk_mem to ensure a new lglkdata global memory section will
be used to create/upgrade the TX log files.

Added Functions:

   None.

Removed Functions:

   None.

CL Interface changes:

   None.

------------------------------------------------------------------------