------------------------------------------------------------------------ r3186 | horda03 | 2010-06-25 13:15:11 -0400 (Fri, 25 Jun 2010) | 89 lines
Cross integration of change 505604

On VMS the Logging/Locking shared memory section is a
Global Section that is deleted when the last session
connected to it drops its connection.

Typically this means when the DMFRCP is stopped the
lglkdata section is released. However, it is possible
for an IPM session to be connected to the lglkdata, if
this is the case then the memory will not be released,
and may still exist when the installation is next started.

The first fix for Bug 122555 handled the above situation
when Ingres was started, INGSTART detects that a stale
lglkdata memory section exists, and defines a logical
(GBL$II_<inst_id>_lglkdata.mem) so that a new lglkdata
memory section will be created.

This change extends the above fix, so that TX log files
can be initialised during an upgrade. If the upgrade
uses a differnet structure for the Logging/Logging system, then
attaching to the memory created by the older version will
cause unexpected errors.

iimkcluster.dcl

Need to know the II_INSTALLATION id, so ensure the value
is obtained (if it exists)

Use check_lglk_mem function (see below) to check the
logging/Locking memory and Online status.

check_lglk_mem returning ii_status_warning implies the
Logging system is online.

iishlib.dcl

Add definition of check_lglk_mem function.

<fixed typo>

check_lglk_mem will return ii_status_ok      if Logging System Offline
ii_status_warning if Online

Define the name of the logical which is used to use an Alternate
lglkdata global memory section.

Do the online check using the memory section that would be in use
(i.e. if alternate area in use, then use that).

Return warning if the Logging system is online.

If the logical name was defined (an system is offline) then extract
the pid from the logical value, and deassign the logical.

Now check if anything is connected to the "default" lglkdata global memory section.
If there is, then define the logical. Note, if the logical was previously
defined, then add 1 to the PID obtained to make a unique number, otherwise
use the PID of the current process.

Return OK as the Logging system is offline.

iisudbms.dcl

Use check_lglk_mem to ensure a new lglkdata global memory section will
be used to create/upgrade the TX log files.

iisunode.dcl

Need to know the II_INSTALLATION id, so ensure the value
is obtained (if it exists)

Use check_lglk_mem to ensure a new lglkdata global memory section will
be used to create/upgrade the TX log files.

Added Functions:

   None.

Removed Functions:

   None.

CL Interface changes:

   None.

------------------------------------------------------------------------
r3187 | loera01 | 2010-06-25 13:30:14 -0400 (Fri, 25 Jun 2010) | 34 lines

This change is a contuation of the fix for bug 122174 introduced in change
498416.  The scope of the change has expanded to SQLGetDescRec(),
SQLColumns(), and SQLDescribeCol().

The bulk of this changes follows up the work begun in change 498416, which
defaults SQL_DESC_LENGTH to the column size rather than the octet length.
In most cases, the column size returns the same value as precision, but
diverge for date/time types that include second fractions:  SQL_DESC_LENGTH is
the number of significant digits in the date, and SQL_DESC_PRECISION is the
number of digits past the decimal point.  For decimal types,
SQL_DESC_PRECISION is the number of significant digits, and SQL_DESC_SCALE
is the number of digits past the decimal point,s one might expect.

Other adjustments were required for SQL_DESC_OCTET_LENGTH.  The ODBC spec
requires SQL_DESC_OCTET_LENGTH to represent the number of bytes required to
convert the column to a character string, as opposed to the number of bytes
in the column as retured from the DBMS.

The Length field in tDESCFLD is unused in ODBC except for metadata queries.
This means that SetDescDefaultsFromType() can set Length directly without
causing regression elsewhere.  Ideally, SetDescDefaultsFromType() would set
all IRD lengths correctly right off the bat.  This is not the case
unfortunately.  For cases in which Length or OctetLength are out of spec,
the appropriate values are set in the metatdata functions themselves.

Change 498416 added a function called GetLengthFromType() to set the Length
Field.  I removed this function and merged its capabilities into
SetDescDefaultsFromType().

I incremented the ODBC version number to 3.50.1000.0120.  This change isn't
associated with a patch, but is significant enough to be associated with a
new version number.

------------------------------------------------------------------------